dead eye almost working + recoil stuff
dead eye works 80 % of the time. sometimes it shoots at all the tags when clicking once, if a bullet hits an enemy before hitting the x in dead eye it doesn't register as a tag hit, need to fix that. Added discouragement from spraying and recoil w deadeye is wip.